export emotion-theming from react-emotion
While refactoring for a proof of concept, it quickly became cumbersome
to have to split up the import just to pull in theming imports.

Since theming functionality is partially built-in by default in emotion v8
via supporting `p => p.theme` etc, I think it makes sense to go the rest
of the way and re-export `emotion-theming` for ease of use.

update snapshots since the file contents changed