Fix "Edit this page" link for docs (#1484)
With #1449, the "Edit this page" link at the top of https://emotion.sh/docs/introduction, https://emotion.sh/docs/class-names, etc are now broken as they direct to `.md` instead of `.mdx`. This PR addresses this issue.

Line 128 does not seem to need `.mdx` because the `README.md` files in `/packages`'s directories have not been converted to `.mdx`.