Add e-prim to the list of libraries
Hello dear team,
I have recently published a library that used in nearly all my projects. It's sort of a system/primitives that's added on top of emotion to make developing new layouts faster. It also contains a few primitive components - box/flex/grid/typography

In the last couple of projects, I've started this effort and evolved it with every different need that I had. I believe the current version is quite usable for most of the projects. The list of supported properties can be extended based on community feedback.