Return cache directly in getServerStylisCache
We are never using the accessor function that is returned.
In the code base later we access it like this:
`serverStylisCache[name]` and not like this `serverStylisCache(name)`

While we could change the access,
I'm not sure how we would write to the cache.
So I think we can just directly return the cache object.