Move creation of stylisCache inside createCache function Currently we share caches even for multiple createCache() calls.
So if you have an SSR app that reates a cache on the server like this:

const cache = createCache({ key: 'custom' });

and you do this (like recommended) on every request, you still share a cache
between all requests. I also think this is currently a memory leak.
If you have a big application and maybe even lots of dynamic styles
the cache can get quite big and will never be garbage collected.